New Header created, small bugfixes
A new header was created for all pages that are not the home page. These pages will not have a splash screen like the home page, so they do not need any of the scrolling functionality. There was also a few bugs present when assigning a few class names which are now all fixed.