Updates per new requirements/feedback from meeting
Various style updates based on the new feedback received in the recent meeting with the GuardianClaims team. Home page had a minor change regarding the text always showing for the claim types, the claims form page has had text added, and small boilerplate text was added to the bottom of the actual claims form. Routing was updated to handle the new pages: "Residential, Commercial, Contractors" and templates were added for these pages which will be implemented next.