Large changes based on GuardianClaims Feedback and meeting
Big style changes were made after gathering feedback from the GuardianClaims team on the home page, about us and claims pages. The templates for the Commercial, Contractors and Residential pages were started, with all of their text content as well as image names and link text (all currently placeholder) added to the content.json file. Currently working on making the "ImageLeftContentRight" and "ImageRightContentLeft" components properly take in the correct content as props. A "WhatWeDo" component was created as well.