Contractor testimonial logos added, content.json updated.
Added contractor testimonial logos and updated the contractor testimonial component to only display these logos. This component is completely dynamic, meaning that you can add or remove logos as you please and the page will handle it automatically. Please refer to the "_comment" found in the .json file for this component for more details/instructions. Also added comments to all dynamic content that is present in the content.json file, this will make it easier for the team to adjust things down the line.