New claims page version
This is the new claims page, as requested by the team at GuardianClaims. All of the text is editable in content.json, which was edited to reflect this new version of the page, claimsContractors in content.json was edited in the same format as well, but the page will not work until this page is finalized and the changes are carried over.